Início implementação Branch-and-Cut-and-Price
Alteração de doubles para floats

+++ VNESolver/Column.h
@@ -13,7 +13,7 @@ class Column {
 		int k, l;
 		int lowerBound;
 		int upperBound;
-		double custoFO;
+		float custoFO;
 
 		Column(int v, int kl);
